fix(review): restructure fleet test helpers and move TestFleetCLIRegistration
- Fix _stub_campaign_resolution to patch autoskillit.cli._fleet.{find_campaign_by_name,
  load_recipe, validate_recipe} — production code uses direct from-imports so source-module
  patches were silently ineffective
- Add DispatchDescriptor NamedTuple to _fleet_helpers.py; update _setup_campaign_with_sessions
  signature from list[tuple[str,str,str]] to list[DispatchDescriptor] with named fields
  (name, status, session_id) to prevent silent position-swap bugs
- Move _get_app/_subcommand_names/_find_command from test_fleet_dispatch.py into
  _fleet_helpers.py alongside all other fleet test utilities
- Move TestFleetCLIRegistration from test_fleet_dispatch.py to test_fleet_list.py —
  registration tests cover the full fleet CLI surface, not dispatch specifically
