+++ tests/cli/test_fleet_list.py
@@ -7,7 +7,12 @@ from pathlib import Path
 import pytest
 
 from autoskillit.cli._fleet import fleet_list as _fleet_list
-from tests.cli._fleet_helpers import _setup_campaign_recipes
+from tests.cli._fleet_helpers import (
+    _find_command,
+    _get_app,
+    _setup_campaign_recipes,
+    _subcommand_names,
+)
 
 pytestmark = [pytest.mark.layer("cli"), pytest.mark.medium, pytest.mark.feature("fleet")]
 
@@ -51,3 +56,45 @@ def test_fleet_list_exits_when_disabled(monkeypatch: pytest.MonkeyPatch, tmp_pat
         _fleet_list()
     assert exc_info.value.code == 1
     assert "fleet" in checked_features
+
+
+# ---------------------------------------------------------------------------
+# CLI registration tests (fleet surface, not list-specific)
+# ---------------------------------------------------------------------------
+
+
+class TestFleetCLIRegistration:
+    def test_fleet_subcommand_registered(self) -> None:
+        app = _get_app()
+        names = _subcommand_names(app)
+        assert "fleet" in names
+
+    def test_fleet_status_accepts_reap_flag(self) -> None:
+        from autoskillit.cli._fleet import fleet_app
+
+        status_cmd = _find_command(fleet_app, "status")
+        assert status_cmd is not None, "fleet status command not found"
+
+    def test_fleet_status_accepts_dry_run_flag(self) -> None:
+        import inspect
+
+        from autoskillit.cli._fleet import fleet_status
+
+        sig = inspect.signature(fleet_status)
+        assert "dry_run" in sig.parameters
+        assert "reap" in sig.parameters
+
+    def test_fleet_dispatch_command_registered(self) -> None:
+        from autoskillit.cli._fleet import fleet_app
+
+        assert "dispatch" in _subcommand_names(fleet_app)
+
+    def test_fleet_campaign_command_registered(self) -> None:
+        from autoskillit.cli._fleet import fleet_app
+
+        assert "campaign" in _subcommand_names(fleet_app)
+
+    def test_fleet_run_command_not_registered(self) -> None:
+        from autoskillit.cli._fleet import fleet_app
+
+        assert "run" not in _subcommand_names(fleet_app)
